refactor: move ApiResponse
to src/common/features/ApiResponse
